requirements.txt
openai-whisper
datasets
transformers
torch
torchvision
facenet-pytorch
mediapipe
pillow
mediapipe 
opencv-python 
numpy
flask
flask-cors
pyannote.audio
python-dotenv
openai